requirements.txt
# --- Core Frameworks ---
fastapi==0.128.8
uvicorn==0.39.0
pydantic==2.12.5
SQLAlchemy==2.0.48

# --- Google AI & API (Loosened to fix the conflict) ---
google-generativeai==0.8.6
google-api-python-client==2.192.0
google-auth==2.49.0
google-api-core[grpc]
grpcio
grpcio-status

# --- Security & Auth ---
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
cryptography==46.0.5
email-validator==2.3.0

# --- Utilities & Tools ---
python-dotenv==1.2.1
python-multipart==0.0.20
httpx==0.28.1
requests==2.32.5
Jinja2==3.1.6
Markdown==3.9
tqdm==4.67.3
click==8.1.8
anyio==4.12.1
certifi==2026.2.25